[json-schema] Add a generic json-schema file for registration and update
